Added Strategy Pattern for matching algorithms
